chore(build): Enable vite to serve from clean repo
No build needed for dev work!
